requirements.txt
tensorflow
numpy
pandas
scikit-learn
Pillow
opencv-python
matplotlib
huggingface_hub
pytorch
torchvision
torchaudio
tqdm
libjpeg-turbo